削除機能に確認機能を追加

+++ src/main/webapp/WEB-INF/jsp/blog.jsp
@@ -38,7 +38,7 @@
 	<p class="c_datetime"><c:out value="${blog.datetime}" /></p>
 	
 	<c:if test="${blog.userId == loginUser.id}">
-		<a href="Delete?id=${blog.id}">削除</a>
+		<a id="delete" href="Delete?id=${blog.id}&page=${paginate.currentPage}" onclick="return confirm('削除しますか？')">削除</a>
 	</c:if>
 	
 	</li>
